requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
olefile==0.44
Pillow==4.3.0
requests==2.18.4
six==1.11.0
slackclient==1.0.9
urllib3==1.22
websocket-client==0.44.0
